@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import {Http ,Response} from '@angular/http';
 import 'rxjs/add/operator/toPromise';
 
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 
